# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tweet Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TweetType = (typeof TweetType)[keyof typeof TweetType];

export const TweetType = {
  text: "text",
  image: "image",
  mixed: "mixed",
  unknown: "unknown",
} as const;

export interface Tweet {
  id: number;
  url: string;
  tweetId: string;
  authorHandle?: string | null;
  authorName?: string | null;
  content?: string | null;
  type: TweetType;
  screenshotUrl?: string | null;
  tags?: string | null;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateTweetBody {
  url: string;
  notes?: string;
  tags?: string;
}

export interface TweetStats {
  total: number;
  textOnly: number;
  imageBased: number;
  mixed: number;
  withScreenshots: number;
  recentlyAdded: number;
}

export interface ErrorResponse {
  error: string;
}

export type ListTweetsParams = {
  type?: ListTweetsType;
  search?: string;
};

export type ListTweetsType =
  (typeof ListTweetsType)[keyof typeof ListTweetsType];

export const ListTweetsType = {
  text: "text",
  image: "image",
  mixed: "mixed",
} as const;