# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tweet Tracker API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: tweets
    description: Tweet tracking operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /tweets:
    get:
      operationId: listTweets
      tags: [tweets]
      summary: List all tracked tweets
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [text, image, mixed]
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of tweets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tweet"
    post:
      operationId: createTweet
      tags: [tweets]
      summary: Add a tweet URL to track
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTweetBody"
      responses:
        "201":
          description: Tweet added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tweet"
        "400":
          description: Invalid URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Tweet already tracked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /tweets/stats:
    get:
      operationId: getTweetStats
      tags: [tweets]
      summary: Get tweet statistics
      responses:
        "200":
          description: Tweet statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TweetStats"

  /tweets/gallery:
    get:
      operationId: getTweetGallery
      tags: [tweets]
      summary: Get tweets with screenshots for gallery view
      responses:
        "200":
          description: Gallery items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tweet"

  /tweets/{id}:
    get:
      operationId: getTweet
      tags: [tweets]
      summary: Get a single tweet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tweet details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tweet"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTweet
      tags: [tweets]
      summary: Delete a tracked tweet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /tweets/{id}/refresh:
    post:
      operationId: refreshTweet
      tags: [tweets]
      summary: Refresh screenshot and metadata for a tweet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Refreshed tweet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tweet"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Tweet:
      type: object
      properties:
        id:
          type: integer
        url:
          type: string
        tweetId:
          type: string
        authorHandle:
          type: string
          nullable: true
        authorName:
          type: string
          nullable: true
        content:
          type: string
          nullable: true
        type:
          type: string
          enum: [text, image, mixed, unknown]
        screenshotUrl:
          type: string
          nullable: true
        tags:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - url
        - tweetId
        - type
        - createdAt
        - updatedAt

    CreateTweetBody:
      type: object
      properties:
        url:
          type: string
        notes:
          type: string
        tags:
          type: string
      required:
        - url

    TweetStats:
      type: object
      properties:
        total:
          type: integer
        textOnly:
          type: integer
        imageBased:
          type: integer
        mixed:
          type: integer
        withScreenshots:
          type: integer
        recentlyAdded:
          type: integer
      required:
        - total
        - textOnly
        - imageBased
        - mixed
        - withScreenshots
        - recentlyAdded

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error