# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tweet Tracker API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all tracked tweets
 */
export const ListTweetsQueryParams = zod.object({
  type: zod.enum(["text", "image", "mixed"]).optional(),
  search: zod.coerce.string().optional(),
});

export const ListTweetsResponseItem = zod.object({
  id: zod.number(),
  url: zod.string(),
  tweetId: zod.string(),
  authorHandle: zod.string().nullish(),
  authorName: zod.string().nullish(),
  content: zod.string().nullish(),
  type: zod.enum(["text", "image", "mixed", "unknown"]),
  screenshotUrl: zod.string().nullish(),
  tags: zod.string().nullish(),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListTweetsResponse = zod.array(ListTweetsResponseItem);

/**
 * @summary Add a tweet URL to track
 */
export const CreateTweetBody = zod.object({
  url: zod.string(),
  notes: zod.string().optional(),
  tags: zod.string().optional(),
});

/**
 * @summary Get tweet statistics
 */
export const GetTweetStatsResponse = zod.object({
  total: zod.number(),
  textOnly: zod.number(),
  imageBased: zod.number(),
  mixed: zod.number(),
  withScreenshots: zod.number(),
  recentlyAdded: zod.number(),
});

/**
 * @summary Get tweets with screenshots for gallery view
 */
export const GetTweetGalleryResponseItem = zod.object({
  id: zod.number(),
  url: zod.string(),
  tweetId: zod.string(),
  authorHandle: zod.string().nullish(),
  authorName: zod.string().nullish(),
  content: zod.string().nullish(),
  type: zod.enum(["text", "image", "mixed", "unknown"]),
  screenshotUrl: zod.string().nullish(),
  tags: zod.string().nullish(),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const GetTweetGalleryResponse = zod.array(GetTweetGalleryResponseItem);

/**
 * @summary Get a single tweet
 */
export const GetTweetParams = zod.object({
  id: zod.coerce.number(),
});

export const GetTweetResponse = zod.object({
  id: zod.number(),
  url: zod.string(),
  tweetId: zod.string(),
  authorHandle: zod.string().nullish(),
  authorName: zod.string().nullish(),
  content: zod.string().nullish(),
  type: zod.enum(["text", "image", "mixed", "unknown"]),
  screenshotUrl: zod.string().nullish(),
  tags: zod.string().nullish(),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a tracked tweet
 */
export const DeleteTweetParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Refresh screenshot and metadata for a tweet
 */
export const RefreshTweetParams = zod.object({
  id: zod.coerce.number(),
});

export const RefreshTweetResponse = zod.object({
  id: zod.number(),
  url: zod.string(),
  tweetId: zod.string(),
  authorHandle: zod.string().nullish(),
  authorName: zod.string().nullish(),
  content: zod.string().nullish(),
  type: zod.enum(["text", "image", "mixed", "unknown"]),
  screenshotUrl: zod.string().nullish(),
  tags: zod.string().nullish(),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});